refactor astronomy.py into base/utils
