+++ ecoscope/analysis/__init__.py
@@ -1,4 +1,4 @@
-from ecoscope.analysis import UD, astronomy, seasons
+from ecoscope.analysis import UD, seasons
 from ecoscope.analysis.ecograph import Ecograph, get_feature_gdf
 from ecoscope.analysis.percentile import get_percentile_area
 from ecoscope.analysis.speed import SpeedDataFrame
@@ -7,7 +7,6 @@ __all__ = [
     "Ecograph",
     "SpeedDataFrame",
     "UD",
-    "astronomy",
     "get_feature_gdf",
     "get_percentile_area",
     "seasons",
